Revert "fix(header): align styles with design"
This reverts commit a96ea2da3c6d16ca4e713108ff1603291f5d4d26.
