+++ apps/rtk-query/src/layout/Header/Header.tsx
@@ -1,15 +1,57 @@
+import { useTranslation } from 'react-i18next'
+
+import { useMeQuery } from '@/features/auth/api'
+import { setIsAuthModalOpen } from '@/features/auth/model'
+import { selectProfileAvatar, selectProfileFullName } from '@/features/profile'
+import { AccountMenu } from '@/layout/Header/AccountMenu'
+import {
+  Button,
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '@/shared/components'
+import { useAppDispatch, useAppSelector } from '@/shared/hooks'
+import { LanguageIcon } from '@/shared/icons/LanguageIcon.tsx'
+import { setLocale } from '@/shared/utils'
+
 import s from './Header.module.css'
-import { HeaderActions } from '@/pages/common/ui/HeaderActions'
 
-type Props = {
-  variant?: 'compact' | 'default'
-}
+export const Header = () => {
+  const { t } = useTranslation()
 
-export const Header = ({ variant = 'default' }: Props) => {
+  const { data: user, isLoading } = useMeQuery()
+  const dispatch = useAppDispatch()
+  const isAuth = !!user
+  const profileAvatarUrl = useAppSelector(selectProfileAvatar)
+  const profileFullName = useAppSelector(selectProfileFullName)
+
   return (
     <header className={s.header}>
       <div className={s.logo}>Musicfun</div>
-      {variant === 'default' && <HeaderActions />}
+      <div className={s.actions}>
+        <DropdownMenu>
+          <DropdownMenuTrigger>
+            <LanguageIcon />
+          </DropdownMenuTrigger>
+          <DropdownMenuContent>
+            <DropdownMenuItem onClick={() => setLocale('en')}>English</DropdownMenuItem>
+            <DropdownMenuItem onClick={() => setLocale('ru')}>Русский</DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+        {isAuth ? (
+          <AccountMenu
+            avatar={profileAvatarUrl}
+            fullName={profileFullName}
+            userLogin={user.login}
+            id={user.userId}
+          />
+        ) : isLoading ? null : (
+          <Button onClick={() => dispatch(setIsAuthModalOpen({ isAuthModalOpen: true }))}>
+            {t('auth.button.sign_in')}
+          </Button>
+        )}
+      </div>
     </header>
   )
 }
