+++ apps/rtk-query/src/pages/PlaylistPage/PlaylistPage.module.css
@@ -1,6 +1,5 @@
 .playlistPage {
   --page-gradient-color: #adbf22;
-  padding-block: 0;
 }
 
 .playlistOverview {
