+++ apps/rtk-query/src/pages/TrackLyricsPage/TrackLyricsPage.module.css
@@ -1,11 +1,8 @@
 .trackLyricsPage {
   --page-gradient-color: #9a3426;
-  padding-block: 0;
 }
 
 .header {
-  display: flex;
-  justify-content: space-between;
   height: var(--header-height);
 }
