+++ apps/rtk-query/src/pages/TrackPage/TrackPage.module.css
@@ -1,6 +1,5 @@
 .trackPage {
   --page-gradient-color: #9a3426;
-  padding-block: 0;
 }
 
 .trackOverview {
