+++ apps/rtk-query/src/pages/TrackPage/TrackPage.tsx
@@ -1,29 +1,30 @@
-import { useTranslation } from 'react-i18next'
-import { useParams } from 'react-router'
+import {useTranslation} from 'react-i18next'
+import {useParams} from 'react-router'
 
-import { useMeQuery } from '@/features/auth'
-import { PlaylistCard, useFetchPlaylistsQuery } from '@/features/playlists'
-import { TrackOverview, useFetchTrackByIdQuery } from '@/features/tracks'
-import { Typography } from '@/shared/components'
-import { ImageType } from '@/shared/types/commonApi.types'
-import { getImageByType } from '@/shared/utils'
+import {useMeQuery} from '@/features/auth'
+import {PlaylistCard, useFetchPlaylistsQuery} from '@/features/playlists'
+import {TrackOverview, useFetchTrackByIdQuery} from '@/features/tracks'
+import {Typography} from '@/shared/components'
+import {ImageType} from '@/shared/types/commonApi.types'
+import {getImageByType} from '@/shared/utils'
 
-import { ContentList, PageWrapper } from '../common'
+import {ContentList, PageWrapper} from '../common'
 import s from './TrackPage.module.css'
-import { ControlPanel } from './ui/ControlPanel'
-import { PageHeader } from '@/pages/common/ui/PageHeader'
+import {ControlPanel} from './ui/ControlPanel'
 
 export const TrackPage = () => {
-  const { t } = useTranslation()
+  const {t} = useTranslation()
 
-  const { id } = useParams()
-  const { data: track } = useFetchTrackByIdQuery({ trackId: id! })
-  const { data: me } = useMeQuery()
+  const {id} = useParams()
+  const {data: track} = useFetchTrackByIdQuery({trackId: id!})
+  const {data: me} = useMeQuery()
   const isTrackOwner = me?.userId === track?.data.attributes.user.id
+
 
   // TODO: backend don't return user id for track
 
-  const { data: playlists } = useFetchPlaylistsQuery({ trackId: id! })
+
+  const {data: playlists} = useFetchPlaylistsQuery({trackId: id!})
 
   if (!track) {
     return <div>{t('tracks.title.tracks_not_found')}</div>
@@ -33,7 +34,6 @@ export const TrackPage = () => {
 
   return (
     <PageWrapper className={s.trackPage}>
-      <PageHeader />
       <TrackOverview
         className={s.trackOverview}
         title={track.data.attributes.title}
@@ -63,6 +63,7 @@ export const TrackPage = () => {
               id={playlist.id}
               title={playlist.attributes.title}
               imageSrc={getImageByType(playlist.attributes.images, ImageType.ORIGINAL)?.url}
+              
             />
           )}
         />
