+++ apps/rtk-query/src/pages/UserPage/UserPage.module.css
@@ -1,7 +1,7 @@
 .userPage {
   --page-gradient-color: #b8a661;
-  padding-top: 0;
 
   display: flex;
   flex-direction: column;
+  gap: 24px;
 }
