+++ apps/rtk-query/src/pages/UserPage/ui/UserInfo/UserInfo.module.css
@@ -3,7 +3,6 @@
   flex-direction: column;
   gap: 16px;
   align-items: center;
-  margin-bottom: 12px;
 }
 
 .descriptionList {
