--- apps/rtk-query/src/pages/common/ui/HeaderActions/HeaderActions.module.css
@@ -1,5 +1,0 @@
-.actions {
-  display: flex;
-  column-gap: 20px;
-  align-items: center;
-}
