--- apps/rtk-query/src/pages/common/ui/HeaderActions/index.ts
@@ -1,1 +1,0 @@
-export * from './HeaderActions.tsx'
