--- apps/rtk-query/src/pages/common/ui/PageHeader/PageHeader.module.css
@@ -1,5 +1,0 @@
-.header {
-  height: var(--header-height);
-  display: flex;
-  justify-content: end;
-}
