--- apps/rtk-query/src/pages/common/ui/PageHeader/PageHeader.tsx
@@ -1,10 +1,0 @@
-import { HeaderActions } from '@/pages/common/ui/HeaderActions'
-import s from './PageHeader.module.css'
-
-export const PageHeader = () => {
-  return (
-    <header className={s.header}>
-      <HeaderActions />
-    </header>
-  )
-}
