--- apps/rtk-query/src/pages/common/ui/PageHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './PageHeader'
