+++ apps/rtk-query/src/shared/components/Tabs/Tabs.module.css
@@ -58,5 +58,5 @@
 }
 
 .tabsContent {
-  padding: 32px 0 0;
+  padding: 32px 0;
 }
