+++ apps/rtk-query/src/shared/icons/index.ts
@@ -33,4 +33,3 @@ export * from './TrackIcon'
 export * from './UploadIcon'
 export * from './VolumeIcon'
 export * from './VolumeMuteIcon'
-export * from './LanguageIcon'
