+++ apps/rtk-query/src/shared/utils/index.ts
@@ -4,4 +4,3 @@ export * from './get-image-by-type'
 export * from './get-user-initials'
 export * from './set-locale'
 export * from './show-error-toast'
-export * from './is-compact-header-path'
